Refactor login.js
